substituted "unsigned int" with "uint16_t"

+++ crypto_sign/cross/clean/keccakf1600.c
@@ -64,18 +64,18 @@ static const uint64_t KeccakF_RoundConstants[NROUNDS] = {
 };
 
 void KeccakF1600_StateExtractBytes(uint64_t *state, unsigned char *data,
-                                   unsigned int offset, unsigned int length)
+                                   uint16_t offset, uint16_t length)
 {
-   unsigned int i;
+   uint16_t i;
    for(i=0; i<length; i++) {
       data[i] = state[(offset + i) >> 3] >> (8*((offset + i) & 0x07));
    }
 }
 
 void KeccakF1600_StateXORBytes(uint64_t *state, const unsigned char *data,
-                               unsigned int offset, unsigned int length)
+                               uint16_t offset, uint16_t length)
 {
-   unsigned int i;
+   uint16_t i;
    for(i = 0; i < length; i++) {
       state[(offset + i) >> 3] ^= (uint64_t)data[i] << (8 * ((offset + i) & 0x07));
    }
